chore(utils): break out utilities.ts into respective files
This file contains a variety of differnt utilities. Slowly moving them to their respective files so they can be properly unit tested and easily extended
