+++ src/pricing.test.ts
@@ -1,5 +1,4 @@
 import {
-  calculateMinimumAuctionBid,
   cloneDemandFactoringData,
   demandFactorPeriodIndex,
   demandIsIncreasing,
@@ -435,32 +434,6 @@ describe('Pricing functions:', () => {
 
         // Ensure input data remains unchanged (weak guarantee due to dependence on cloneDemandFactoringData)
         expect(inputDfData).toEqual(clonedInputDfData);
-      },
-    );
-  });
-
-  describe('calculateMinimumAuctionBid function', () => {
-    it.each([
-      [[0, 0, 0.1], 100],
-      [[0, 1, 0.1], 90],
-      [[0, 2, 0.1], 81],
-      [[0, 3, 0.1], 72.9],
-      [[0, 0, 0.2], 100],
-      [[0, 1, 0.2], 80],
-      [[0, 2, 0.2], 64],
-      [[0, 3, 0.2], 51.2],
-    ])(
-      'given [current block height, moving average purchase count] of %j, should return %d',
-      ([startHeight, currentHeight, decayRate], expectedMinimumBid) => {
-        const calculatedMinimumBid = calculateMinimumAuctionBid({
-          startHeight: new BlockHeight(startHeight),
-          startPrice: 100,
-          floorPrice: 10,
-          currentBlockHeight: new BlockHeight(currentHeight),
-          decayRate: decayRate, // 10% per interval
-          decayInterval: 1,
-        });
-        expect(calculatedMinimumBid.valueOf()).toEqual(expectedMinimumBid);
       },
     );
   });
